# resources/Wasteless_API_Spec.yml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: >-
      https://virtserver.swaggerhub.com/matthewminish/seng302-2021-api-spec/3.0.1
info:
  description: API specification for the SENG302 2021 project team 800
  version: 6.0.0
  title: Wasteless
  contact:
    email: seng302@canterbury.ac.nz
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  /cards:
    post:
      tags:
        - 'UCM2: Card creation'
        - Cards / community marketplace
      summary: Create a new card to go on the community marketplace
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '201':
          description: Card created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  cardId:
                    type: integer
                    description: The ID of the card that has just been created
                    example: 500
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: >-
            Returned if a user who is not a GAA tries to create a card for
            somebody else.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Card'
    get:
      tags:
        - 'UCM2: Card creation'
        - Cards / community marketplace
      summary: Retrieve all cards for a given section
      parameters:
        - name: section
          in: query
          description: The section for which cards should be retrieved.
          required: true
          schema:
            type: string
            enum:
              - ForSale
              - Wanted
              - Exchange
        - name: page
          in: query
          description: The page of that section should be retrieved
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          description: The number of items per page
          schema:
            type: integer
            minimum: 1
            default: 8
        - name: sortBy
          in: query
          description: Sort the paginated cards by a given element
          schema:
            type: string
            enum:
              - DATE_ASC
              - DATE_DESC
              - TITLE_AZ
              - TITLE_ZA
        - name: country
          in: query
          description: Filter by a specific country
          schema:
            type: string
        - name: region
          in: query
          description: Filter by a region (must also have country!)
          schema:
            type: string
        - name: city
          in: query
          description: Filter by a city (must also have region!)
          schema:
            type: string
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: Request successful, a (possibly empty) list of cards is returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                    minimum: 1
                    default: 1
                  totalElements:
                    type: integer
                    minimum: 1
                    default: 8
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /cards/{id}:
    get:
      tags:
        - 'UCM3: Marketplace section display'
        - Cards / community marketplace
      summary: Retrieve a specific card
      parameters:
        - name: id
          in: path
          description: The id of the card to get.
          required: true
          schema:
            type: integer
            example: 500
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: Request successful, a card is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    delete:
      tags:
        - 'UCM4: Card expiry'
        - Cards / community marketplace
      summary: Delete a community marketplace card
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: The id of the card to get.
          required: true
          schema:
            type: integer
            example: 500
      responses:
        '200':
          description: Card deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: >-
            Forbidden: Returned when a user tries to delete a card that they are
            not the creator of AND the user is not a GAA
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    put:
      tags:
        - 'UCM9: Edit card'
        - Cards / community marketplace
      summary: Edit a community marketplace card
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdim
      parameters:
        - name: id
          in: path
          description: The id of the card to get.
          required: true
          schema:
            type: integer
            example: 500
      responses:
        '200':
          description: Card information has been updated correctly
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: >-
            Forbidden: Returned when a user who is neither the creator of the
            card, nor a GAA, tries to extend the display period of the card.
        '404':
          description: Card not found
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Card'
        description: Updated card data
  /cards/{id}/extenddisplayperiod:
    put:
      tags:
        - 'UCM4: Card expiry'
        - Cards / community marketplace
      summary: Extend the display period of a card that is nearing expiry
      parameters:
        - name: id
          in: path
          description: The id of the card to get.
          required: true
          schema:
            type: integer
            example: 500
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: >-
            End of display period successfully extended. Extension amount should
            be a further two weeks.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: >-
            Forbidden: Returned when a user who is neither the creator of the
            card, nor a GAA, tries to extend the display period of the card.
        '406':
          $ref: '#/components/responses/NotAcceptableError'
...
  /users/{id}/cards:
    get:
      tags:
        - 'UCM5: Find my cards'
        - Cards / community marketplace
      summary: Retrieve all active cards that a user has created
      parameters:
        - name: id
          in: path
          description: The id of the user for whom to get cards.
          required: true
          schema:
            type: integer
            example: 100
        - name: page
          in: query
          description: The page of that section should be retrieved
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          description: The number of items per page
          schema:
            type: integer
            minimum: 1
            default: 6
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: Request successful, a (possibly empty) list of cards is returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                    minimum: 1
                    default: 1
                  totalElements:
                    type: integer
                    minimum: 1
                    default: 6
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
...
  /cards/search:
    get:
      tags:
        - 'UCM7: Keyword search'
        - Cards / community marketplace
        - Keywords
      summary: Search for cards based on keywords and location attributes
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - in: query
          name: keywords
          description: a list of keyword strings
          schema:
            type: array
            items:
              type: string
          required: true
        - in: query
          name: type
          description: >-
            Whether or not to get cards that match ANY of the keywords, or only
            cards that match ALL of the keywords. Or = Match ANY, And = Match
            ALL
          schema:
            type: string
            enum:
              - And
              - Or
            default: And
            required: true
        - name: section
          in: query
          description: The section for which cards should be retrieved.
          required: true
          schema:
            type: string
            enum:
              - ForSale
              - Wanted
              - Exchange
        - in: query
          name: Country
          description: The country to filter cards by
          schema:
            type: string
            example: New Zealand, Australia, United States...
          required: true
        - in: query
          name: Region
          description: The region to filter cards by
          schema:
            type: string
            example: Northland, Canterbury...
          required: true
        - in: query
          name: City
          description: The city to filter cards by
          schema:
            type: string
            example: Auckland, Melbourne....
          required: true
        - name: page
          in: query
          description: The page of that section should be retrieved
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          description: The number of items per page
          schema:
            type: integer
            minimum: 1
            default: 8
      responses:
        '200':
          description: Request successful, a (possibly empty) list of cards is returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                    minimum: 1
                    default: 1
                  totalElements:
                    type: integer
                    minimum: 1
                    default: 10
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    BadRequestError:
      description: >-
        There was some error with the data supplied by the user, appropriate
        error message(s) should be shown to user
    UnauthorizedError:
      description: Access token is missing or invalid
...
    NotAcceptableError:
      description: >-
        The requested route does exist (so not a 404) but some part of the
        request is not acceptable, for example trying to access a resource by an
        ID that does not exist.
...
  schemas:
...
    Keyword:
      type: object
      description: Represents a single 'keyword' within the system - similar to a hashtag.
      properties:
        id:
          type: integer
          example: 600
          readOnly: true
          description: Unique ID, autogenerated by the system
        name:
          type: string
          description: The text value of the keyword
          example: Vehicle
        created:
          type: string
          format: date-time
          example: '2021-07-15T05:10:00Z'
          readOnly: true
          description: Date and time when the keyword was created
    Card:
      type: object
      description: Represents an item in the Community Marketplace
      required:
        - title
        - creatorId
        - section
        - keywordIds
      properties:
        id:
          type: integer
          example: 500
          readOnly: true
          description: Unique ID, autogenerated by the system
        creatorId:
          type: integer
          example: 100
          writeOnly: true
        creator:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/User'
        section:
          type: string
          enum:
            - ForSale
            - Wanted
            - Exchange
        created:
          type: string
          format: date-time
          example: '2021-07-15T05:10:00Z'
          readOnly: true
          description: Date and time when the card was created
        displayPeriodEnd:
          type: string
          format: date-time
          example: '2021-07-29T05:10:00Z'
          readOnly: true
          description: >-
            Date and time when the display period of the card is set to expire.
            This defaults to 2 weeks from creation, and can be extended per
            UCM4.
        title:
          type: string
          description: Title of the card, should be suitable for one-line displays.
          example: 1982 Lada Samara
        description:
          type: string
          description: A more detailed description of the card.
          example: >-
            Beige, suitable for a hen house. Fair condition. Some rust. As is,
            where is. Will swap for budgerigar.
        keywordsObj:
          type: array
          readOnly: true
          description: >-
            Keywords describing the card. A card can have one or more keywords
            associated with it.
          items:
            $ref: '#/components/schemas/Keyword'
        keywords:
          type: array
          writeOnly: true
          description: >-
            The names of keywords that will be associated with a card once it is
            created. These keywords are created through the POST /keywords
            endpoints (UCM6).
          items:
            type: integer
            example:
              - Car
              - Auto
              - Vehicle
...
    Business:
      type: object
      description: Business account entity
      required:
        - name
        - address
        - businessType
      properties:
        id:
          type: integer
          example: 100
          readOnly: true
          description: Automatically generated and assigned by the server
        administrators:
          type: array
          items:
            $ref: '#/components/schemas/User'
          readOnly: true
        primaryAdministratorId:
          type: integer
          example: 20
          description: >-
            The ID of the 'primary' administrator of the business. This is
            automatically given to the creator of the business.
        name:
          type: string
          example: Lumbridge General Store
        description:
          type: string
          example: A one-stop shop for all your adventuring needs
        address:
          $ref: '#/components/schemas/Address'
        businessType:
          type: string
          enum:
            - Accommodation and Food Services
            - Retail Trade
            - Charitable organisation
            - Non-profit organisation
        created:
          type: string
          format: date-time
          example: '2020-07-14T14:52:00Z'
          readOnly: true
          description: Date and time when the account was created
    Address:
      type: object
      description: Entity representing some physical address
      required:
        - country
      properties:
        streetNumber:
          type: string
          example: 3/24
          writeOnly: true
        streetName:
          type: string
          example: Ilam Road
          writeOnly: true
        suburb:
          type: string
          example: Upper Riccarton
          writeOnly: true
        city:
          type: string
          example: Christchurch
        region:
          type: string
          example: Canterbury
        country:
          type: string
          example: New Zealand
        postcode:
          type: string
          example: '90210'
          writeOnly: true
    User:
      type: object
      description: Basic user entity
      required:
        - firstName
        - lastName
        - email
        - dateOfBirth
        - homeAddress
        - password
      properties:
        id:
          type: integer
          example: 100
          readOnly: true
          description: Automatically generated and assigned by the server
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Smith
        middleName:
          type: string
          example: Hector
        nickname:
          type: string
          example: Jonny
        bio:
          type: string
          example: Likes long walks on the beach
        email:
          type: string
          format: email
          example: johnsmith99@gmail.com
          writeOnly: true
        primaryImageId:
          type: integer
          example: 1
          description: >-
            The ID of the 'primary' image chosen by the user. By default it is
            the first image the use uploads.
        countryForCurrency:
          type: string
          description: >-
            The country in which the currency is used, if not initialized then
            default to New Zealand
          example: New Zealand
        dateOfBirth:
          type: string
          format: date
          example: '1999-04-27'
          writeOnly: true
        phoneNumber:
          type: string
          example: +64 3 555 0129
          writeOnly: true
        homeAddress:
          $ref: '#/components/schemas/Address'
        password:
          type: string
          example: 1337-H%nt3r2
          format: password
          writeOnly: true
          description: Should not be stored in plain text (i.e hash it)
        created:
          type: string
          format: date-time
          example: '2020-07-14T14:32:00Z'
          readOnly: true
          description: Date and time when the account was created
        role:
          type: string
          readOnly: true
          description: >-
            (Implemented in U4): This property should only be shown to Global
            Application Admins
          enum:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
        businessesAdministered:
          type: array
          readOnly: true
          description: >-
            (Implemented in U6): A list of businesses for which the user is an
            administrator
          items:
            $ref: '#/components/schemas/Business'
...